+++ src/Applications/GAAS_App/viirs_l2a.py
@@ -46,11 +46,47 @@ if __name__ == "__main__":
                   " --fname=" + cf('VIIRS_L2A_OUT_TEMPLATE') + \
                     " --net=" + cf('VIIRS_L2A_NN_FILE')      + \
                  " --aer_x="  + cf('VIIRS_L2A_AER_X')        + \
-              " --blank_ods=" + cf('VIIRS_L2A_BLANK_ODS')   
+              " --blank_ods=" + cf('VIIRS_L2A_BLANK_ODS')     
 
     if cf('VIIRS_L2A_OVERWRITE').upper() == 'YES': Options += " --force"
     if   cf('VIIRS_L2A_VERBOSE').upper() == 'YES': Options += " -v"
-              
+
+    try:
+        cloud_thresh = cf('VIIRS_L2A_CLOUD_THRESH')
+        Options += " --cloud_thresh=" + cloud_thresh
+    except:
+        pass
+
+    try:
+        cloudFree = cf('VIIRS_L2A_CLOUDFREE')
+        Options += " --cloudFree=" + cloudFree
+    except:
+        pass
+
+    try:
+        aodmax = cf('VIIRS_L2A_AODMAX')
+        Options += " --aodmax=" + aodmax
+    except:
+        pass
+
+    try:
+        aodSTD = cf('VIIRS_L2A_AODSTD')
+        Options += " --aodSTD=" + aodSTD
+    except:
+        pass
+
+    try:
+        aodLenth = cf('VIIRS_L2A_AODLENGTH')
+        Options += " --aodLength=" + aodLength
+    except:
+        pass
+
+    try:
+        wavs = cf('VIIRS_L2A_WAVS')
+        Options += " --wavs=" + wavs
+    except:
+        pass    
+
     # Generate products
     # -----------------
     i = 0