+++ src/Applications/GAAS_App/vx04_l2a.py
@@ -42,7 +42,8 @@ def makethis_dir(filename):
             raise IOError("could not create directory "+path)
         
 #---------------------------------------------------------------------
-
+def wavs_callback(option, opt, value, parser):
+    setattr(parser.values, option.dest, value.split(','))
 if __name__ == "__main__":
 
     expid = 'nnr_001'
@@ -67,6 +68,12 @@ if __name__ == "__main__":
     coll = '002'
     res = 'c'
     nsyn = 8
+    cloud_thresh = 0.70
+    cloudFree = None
+    aodmax = 1.0
+    aodSTD = 3.0
+    aodLength = 0.5
+    wavs = '440,470,550,660,870'
     
 #   Parse command line options
 #   --------------------------
@@ -112,7 +119,31 @@ if __name__ == "__main__":
 
     parser.add_option("-r", "--res", dest="res", default=res,
                       help="Resolution for gridded output (default=%s)"\
-                           %out_tmpl )
+                           %res )
+
+    parser.add_option("--cloud_thresh", dest="cloud_thresh", default=cloud_thresh,type='float',
+                      help="Cloud fractions threshhold for good data (default=%f)"\
+                           %cloud_thresh )    
+
+    parser.add_option("--cloudFree", dest="cloudFree", default=cloudFree,
+                      help="Extra check for cloudiness when high AOD values are predicted. If not provided, no check is performed. (default=%s)"\
+                           %cloudFree )
+
+    parser.add_option("--aodmax", dest="aodmax", default=aodmax,type='float',
+                      help="max AOD value that will be accepted when cloud fraction is greater than cloudFree (default=%f)"\
+                           %aodmax )
+
+    parser.add_option("--aodSTD", dest="aodSTD", default=aodSTD,type='float',
+                      help="number of standard deviations to check for AOD outliers (default=%f)"\
+                           %aodSTD )
+
+    parser.add_option("--aodLength", dest="aodLength", default=aodLength,type='float',
+                      help="length scale (degrees) to check for AOD outliers (default=%f)"\
+                           %aodLength )
+
+    parser.add_option("--wavs", dest="wavs", default=wavs,type='string',action='callback',callback=wavs_callback,
+                      help="wavelength to output AOD from predicted Angstrom Exponent (default=%s)"\
+                           %wavs )
 
     parser.add_option("-u", "--uncompressed",
                       action="store_true", dest="uncompressed",default=False,
@@ -188,11 +219,21 @@ if __name__ == "__main__":
     if options.verbose:
         print("NNR Retrieving %s %s on "%(sat,algo.upper()),syn_time)
 
+    if options.cloudFree == 'None':
+        options.cloudFree = None
+    elif options.cloudFree == None:
+        pass        
+    else:
+        options.cloudFree = float(options.cloudFree)
+
     viirs = Vx04_NNR(options.l2_path,sat,algo.upper(),syn_time,aer_x,
                       coll=options.coll,
-                      cloud_thresh=0.7,
-                      cloudFree = 0.0,
-                      aodmax = 1.0,
+                      cloud_thresh=options.cloud_thresh,
+                      cloudFree=options.cloudFree,
+                      aodmax=options.aodmax,
+                      aodSTD = options.aodSTD,
+                      aodLength = options.aodLength,
+                      wavs = options.wavs,                      
                       nsyn=options.nsyn,                      
                       verbose=options.verbose)
     if viirs.nobs < 1: